chore(react-tooltip): revert redundant base hook
