+++ packages/react-components/react-tooltip/library/src/Tooltip.ts
@@ -1,10 +1,8 @@
 export type {
   OnVisibleChangeData,
   TooltipChildProps,
-  TooltipBaseProps,
   TooltipProps,
   TooltipSlots,
-  TooltipBaseState,
   TooltipState,
 } from './components/Tooltip/index';
 export {
@@ -13,5 +11,4 @@ export {
   tooltipClassNames,
   useTooltipStyles_unstable,
   useTooltip_unstable,
-  useTooltipBase_unstable,
 } from './components/Tooltip/index';
