+++ packages/react-components/react-tooltip/library/src/components/Tooltip/Tooltip.types.ts
@@ -122,8 +122,6 @@ export type TooltipProps = ComponentProps<TooltipSlots> &
     withArrow?: boolean;
   };
 
-export type TooltipBaseProps = Omit<TooltipProps, 'appearance'>;
-
 /**
  * State used in rendering Tooltip
  */
@@ -147,5 +145,3 @@ export type TooltipState = ComponentState<TooltipSlots> &
      */
     arrowClassName?: string;
   };
-
-export type TooltipBaseState = Omit<TooltipState, 'appearance'>;
