+++ packages/react-components/react-tooltip/library/src/components/Tooltip/index.ts
@@ -1,14 +1,5 @@
 export { Tooltip } from './Tooltip';
-export type {
-  OnVisibleChangeData,
-  TooltipChildProps,
-  TooltipBaseProps,
-  TooltipProps,
-  TooltipSlots,
-  TooltipBaseState,
-  TooltipState,
-} from './Tooltip.types';
+export type { OnVisibleChangeData, TooltipChildProps, TooltipProps, TooltipSlots, TooltipState } from './Tooltip.types';
 export { renderTooltip_unstable } from './renderTooltip';
 export { useTooltip_unstable } from './useTooltip';
-export { useTooltipBase_unstable } from './useTooltipBase';
 export { tooltipClassNames, useTooltipStyles_unstable } from './useTooltipStyles.styles';
