+++ packages/react-components/react-tooltip/library/src/components/Tooltip/renderTooltip.tsx
@@ -4,12 +4,12 @@
 import { Portal } from '@fluentui/react-portal';
 import { assertSlots } from '@fluentui/react-utilities';
 import type { JSXElement } from '@fluentui/react-utilities';
-import type { TooltipBaseState, TooltipSlots } from './Tooltip.types';
+import type { TooltipState, TooltipSlots } from './Tooltip.types';
 
 /**
  * Render the final JSX of Tooltip
  */
-export const renderTooltip_unstable = (state: TooltipBaseState): JSXElement => {
+export const renderTooltip_unstable = (state: TooltipState): JSXElement => {
   assertSlots<TooltipSlots>(state);
 
   return (
