+++ packages/react-components/react-tooltip/library/src/index.ts
@@ -4,7 +4,6 @@ export {
   tooltipClassNames,
   useTooltipStyles_unstable,
   useTooltip_unstable,
-  useTooltipBase_unstable,
 } from './Tooltip';
 export type {
   OnVisibleChangeData,
@@ -12,6 +11,4 @@ export type {
   TooltipSlots,
   TooltipState,
   TooltipChildProps as TooltipTriggerProps,
-  TooltipBaseProps,
-  TooltipBaseState,
 } from './Tooltip';
